+++ modules/hanna-react/src/utils.ts
@@ -6,3 +6,12 @@ export * from './utils/useFormatMonitor.js';
 export * from './utils/useGetSVGtext.js';
 export * from './utils/useMixedControlState.js';
 export * from './utils/useScrollbarWidthCSSVar.js';
+
+/**
+ * Helper type to add HTML element props to a component, **including**
+ * `data-*` attributes
+ */
+export type HTMLProps<T extends keyof JSX.IntrinsicElements> =
+  JSX.IntrinsicElements[T] & {
+    [dataAttr: `data-${string}`]: unknown;
+  };